WebSockets/tests/submissions/Opera/interfaces/WebSocket/send/001.html

<!doctype html>
<title>WebSockets: send() with no args</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
  assert_throws(new TypeError(), function(){ws.send()});
});
</script>

